fix(sorting): Sort on housenumber when distance tied
The interpolation code relies on a particular ordering of address
records: sorted by distance from the start of a street, and then ideally
sorted consistently by address.

The sorting code was sorting only by distance, which could lead to
differences in the resulting array sorting depending on the Node.js
version.

This PR changes the sorting so that it explicitly considers the house
number in the case of ties on distance.