Move documentation from Pelias Wiki to this repo
In order to allow us to deprecate the pelias/pelias wiki, this moves all
interpolation related documentation to this repository.

GitHub wikis are a bit of a niche feature, which means any information
in there is hard to find.

Additionally, because they don't notify us when changes are made (in
comparison to the normal pull request workflow of documentation stored
in a git repositiory proper), we have no idea what the state of anything
is.