CI: use Ubuntu 18 for testing
Test output can depend on the version of `sqlite3` installed, and
currently our tests expect the output from roughly sqlite3 `v3.22`,
which is what comes with Ubuntu 18.

Eventually we'll upgrade everything to Ubuntu 20, but for now it's best
to stick with the version of Ubuntu used in our Docker baseimage.